src/DTW_KNN/requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
blessings==1.7
brotlipy==0.7.0
cachetools==4.2.1
certifi==2020.6.20
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Cython==0.29.23
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
docopt==0.6.2
dtaidistance==2.2.5
entrypoints==0.3
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.14
google-api-core==1.31.1
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
grpcio==1.32.0
h5py==2.10.0
hiredis==2.0.0
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonpickle==1.5.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Lasagne==0.1
libpython==0.2
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
msgpack==1.0.2
multidict==5.1.0
multiprocess==0.70.12.2
munch==2.5.0
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numba==0.53.1
numpy==1.18.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus==0.7.13
opencensus-context==0.1.2
opt-einsum==3.3.0
p-tqdm==1.3.3
packaging==20.9
pandas==1.0.1
pandocfilters==1.4.3
parso==0.8.2
pathos==0.2.8
pickleshare==0.7.5
Pillow==8.2.0
pox==0.3.0
ppft==1.6.6.4
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.15.6
psutil==5.8.0
py-cpuinfo==7.0.0
py-spy==0.3.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pydantic==1.8.2
Pygments==2.9.0
pymongo==3.11.4
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
ray==1.5.1
redis==3.5.3
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel-yaml==0.15.87
sacred==0.8.2
scikit-learn==0.24.1
scipy==1.4.1
Send2Trash==1.7.1
smmap==3.0.5
SQLAlchemy==1.3.13
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
Theano==1.0.5
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
tslearn==0.5.1.0
typing-extensions==3.7.4.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
yarl==1.6.3
